requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pyasn1==0.6.1
pycparser==2.22
pysmb==1.2.10
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.1
Werkzeug==3.1.3
gunicorn==23.0.0
